Move `hasNewlineBetweenOrAfterDecorators`
